requirements.txt
urllib3==1.26.12
py-cord==2.2.2
requests==2.28.1
python-dotenv==0.21.0